Make an attempt to parse the response body as XML if the user is expecting XML, even if the Content-Type says otherwise.
Prior to this patch, if an XML document is sent with the wrong Content-Type (e.g. "text/html"), jQuery will return a parsererror, even if the response body is valid XML. This issue was discovered via a Rails bug in which cached content (in our case, an XML document) is returned with a Content-Type of "text/html" rather than the original Content-Type. See https://rails.lighthouseapp.com/projects/8994/tickets/1585-action-caching-sets-wrong-content-type-when-cache_path-is-a-string

Ticket #5764 (http://dev.jquery.com/ticket/5764)